3d textures working.

+++ VulkanRayTraceTest/SceneData.h
@@ -31,7 +31,7 @@ struct SceneDataBufferData {
 	alignas(4) float vertexSize;
 	alignas(16) glm::mat4 PVM;
 	alignas(16) glm::mat4 BoneTransform[100];
-	alignas(4) int DepthSampler;
+	alignas(4) float DepthSampler;
 };
 
 struct SceneDataStruct
